style: Normal font sizes, and use monospace for listings
There's a significant amount of overrides to make the font sizes
smaller, but that can hurt readability in some cases. We should try to
use the "natural" sizes as much as possible.

This patch does that, removing a lot of the font sizes and bringing them
to be based on the normal sizes.

It also changes listings to use monospace, for readability.