Add basic submodules support
Currently, submodule paths are skipped in the tree list.

That makes them invisible, which is quite inconvenient.

This patch makes git-arr show submodule entries in the listing, and
mentions they are a submodule. This is the bare minimum support
possible, but it is much better than skipping the entries.

Thanks to Stephen Paul Weber <singpolyma@singpolyma.net> for noticing
the issue and suggesting this improvement!