Add optional debug tracing to help with performance analysis
It can be very time consuming to identify why git-arr is taking time to
run.

While end users are not expected to do that kind of debugging, it can be
very useful for specific situations.

Today, git-arr performance is quite opaque and it's hard to know what is
it doing or how long different calls are taking.

This patch introduces some basic tracing, that is shown to stderr when
the GIT_ARR_DEBUG env variable is set.

This is a debugging aid and subject to change in the future in
backwards-incompatible ways.