Skip calling write_to() more than once for the same commit
The commit view does not depend on the branch, and the write_to()
function will skip the write if the output file already exists.

However, that existance check still takes a significant amount of time,
especially when extending an existing repo.

So to speed up re-generation of static content, this patch makes git-arr
keep track of commits that we've already generated.

On a repo with 4 branches and 600 commits, this results in ~10% speedup
on a no-op regeneration.