Only call mkdir if we are writing the file
Today, when generating files, we unconditionally check if the directory
exists, and create it if not.

However, most of the time when re-generating we skip writing the file
altogether, which results in a lot of unnecessary checks.

This patch moves the directory creation right next to the file writing,
which results in a small speedup while regenerating.