Speed up finding the git directory for a repository
When loading the config, we need to find the location of the git
directory for all repositories.

We do this even if we are only operating on a single repo (because we
need to as part of the config, and also to generate the top level index).

Today, this is implemented by calling `git rev-parse --git-dir`;
however, that ends up taking a relatively significant amount of time
when regenerating a single repo.

To speed things up, this patch introduces a simpler heuristic that
should be good enough for most purposes. We expect the paths to be
repositories already.

On a config with 34 repos, a no-op regeneration of a repo with 4 branches and
600 commits, this results in ~15% speedup.