git: Make smstr.html and smstr.url cached properties
We use smstr a lot when parsing git output, but a lot of times we don't
need the HTML or URL conversions. However, today we always compute them.

This patch makes smstr.html and smstr.url cached properties, so they
will only be computed when needed.

This results in a noticeable speedup when regenerating repositories.