Use xattr to avoid rewriting all blobs on each commit
Today, on each commit we regenerate the entire tree for the relevant
branch, which takes a significant amount of time.

This patch records the object id of each blob when it is written, and
then avoids any later writes if the object id is the same.

The result is a significant performance improvement on incremental
regeneration.

If the python3-xattr package is not available, we fall back to the
previous behaviour of using the mtime of the branch.