Use a separate file for repositories' "last modified time"
Today, when regenerating a single repositoriy (with --only), we spend
most of the time going through the other repositories and finding their
last modified time, just to put it in the top-level index.

To speed that up, we stop putting the timestamp inline in the top-level
index, and instead have a separate modified_ts.json file to keep the
timestamp for each repo.

Then, when using --only, we update that file with the data from the
repositories being updated only, instead of all of them.

The resulting speedup depends on the number of repositories, but on a 40
repo config, when using --only this results in a 1s -> 0.5s improvement.